Use long also internally
This should have been part of the commit 5c5d66a6a69002e1e89f433000b143f6310c2282

+++ src/main/java/ro/gal/perfectnumber/service/PerfectNumberService.java
@@ -18,7 +18,7 @@ public class PerfectNumberService {
             return false;
         }
         int sumOfDividers = 1;
-        for (int i=2; i*i<=number; i++) {
+        for (long i=2; i*i<=number; i++) {
             if (number%i == 0) {
                 sumOfDividers += i;//add divisor
                 sumOfDividers += number / i;//add the other divisor
